update 6 files, delete 5 files, create 1 file and move 2 files
