--- src/components/Category/Category.astro
@@ -1,17 +1,0 @@
----
-type Props = {
-  category: string;
-};
-const { category } = Astro.props;
-const categoryColors: { [key: string]: string } = {
-  Tech: 'text-content',
-  Project: 'text-accent',
-  Diary: 'text-green-700 dark:text-green-300',
-};
-
-const textColorClass = categoryColors[category];
----
-
-<span class=`w-fit ${textColorClass}`>
-  {category}
-</span>
