+++ src/components/Header/Logo.astro
@@ -5,7 +5,7 @@ const siteName = siteConfig.site.title;
 
 <a
   href="/"
-  class="flex place-items-center text-lg font-bold text-content px-1"
+  class="flex place-items-center px-1 text-lg font-bold text-content"
   aria-label="Go to Homepage"
 >
   <span>{siteName}</span>
