+++ src/components/Home/LatestPosts.astro
@@ -1,19 +1,19 @@
 ---
-import PostList from '@components/PostList/PostList.astro';
+import PostList from '@components/Posts/PostList.astro';
 import { sortPostsByDateDesc } from '@utils/data-util';
 import { type CollectionEntry, getCollection } from 'astro:content';
 
 const LATEST_POSTS_SIZE = 4;
 const TITLE = '최신 글';
 
-const latestPosts: CollectionEntry<'blog'>[] = (
-  await getCollection('blog')
-).sort(sortPostsByDateDesc);
+const latestPosts: CollectionEntry<'blog'>[] = (await getCollection('blog'))
+  .sort(sortPostsByDateDesc)
+  .slice(0, LATEST_POSTS_SIZE);
 ---
 
 <section class="flex flex-col space-y-4">
   <span class="pb-2 text-3xl font-black text-content">
     {TITLE}
   </span>
-  <PostList pageSize={LATEST_POSTS_SIZE} posts={latestPosts} />
+  <PostList posts={latestPosts} />
 </section>
