--- src/components/PostCard/Description.astro
@@ -1,8 +1,0 @@
----
-type Props = { description: string };
-const { description } = Astro.props;
----
-
-<p class="line-clamp-2 break-keep text-sm text-subContent/90 sm:text-base">
-  {description}
-</p>
