--- src/components/PostCard/Header.astro
@@ -1,18 +1,0 @@
----
-import Category from '@components/Category/Category.astro';
-import FormattedDate from '@components/Date/FormattedDate.astro';
-
-type Props = {
-  category: string;
-  publishedDate: Date;
-};
-const { category, publishedDate } = Astro.props;
----
-
-<div
-  class="flex items-center gap-x-2 text-xs font-medium text-subContent sm:text-sm"
->
-  <Category category={category} />
-  <span>|</span>
-  <FormattedDate date={publishedDate} />
-</div>
