--- src/components/PostCard/Post.astro
@@ -1,26 +1,0 @@
----
-import Header from './Header.astro';
-import Title from './Title.astro';
-import Description from './Description.astro';
-
-type Props = {
-  title: string;
-  description: string;
-  publishedDate: Date;
-  link: string;
-  category: string;
-};
-
-const { title, description, publishedDate, link, category } = Astro.props;
----
-
-<a href={link} class="flex flex-col">
-  <article
-    aria-label={title}
-    class="group relative w-full overflow-hidden break-keep rounded-lg transition-all duration-300 ease-in-out hover:animate-moveRight hover:bg-gray-100/90 hover:dark:bg-gray-700/90 px-2 py-1"
-  >
-    <Header category={category} publishedDate={publishedDate} />
-    <Title title={title} />
-    <Description description={description} />
-  </article>
-</a>
