--- src/components/PostCard/Title.astro
@@ -1,12 +1,0 @@
----
-type Props = {
-  title: string;
-};
-const { title } = Astro.props;
----
-
-<h2
-  class="text-lg font-medium text-content transition-all ease-in-out hover:text-accent lg:text-2xl"
->
-  {title}
-</h2>
