+++ src/components/Posts/Card.astro
@@ -1,0 +1,24 @@
+---
+type Props = {
+  title: string;
+  description: string;
+  link: string;
+};
+
+const { title, description, link } = Astro.props;
+---
+
+<a href={link} aria-label={title}>
+  <article
+    class="group relative w-full overflow-hidden break-keep rounded-lg p-2 transition-all duration-300 ease-in-out hover:animate-moveRight hover:bg-gray-100/90 hover:dark:bg-gray-700/90"
+  >
+    <h2
+      class="text-lg font-medium text-content transition-all ease-in-out hover:text-accent lg:text-2xl"
+    >
+      {title}
+    </h2>
+    <p class="line-clamp-2 break-keep text-sm text-subContent/90 sm:text-base">
+      {description}
+    </p>
+  </article>
+</a>
