+++ src/layouts/BaseHead.astro
@@ -1,5 +1,4 @@
 ---
-
 import SpeedInsights from '@vercel/speed-insights/astro';
 import { ViewTransitions } from 'astro:transitions';
